Answer of isAdult
Given the age of a person and the adulthood age in a particular country, determine if the person is considered an adult in that country. Note that people whose age is the same as the adulthood age are considered adult.

Example

For adulthoodAge = 21 and age=18, the answer should be false.

[input] integer age

The age of the person, a positive integer.
[input] integer adulthoodAge

The adulthood age, a positive integer.
[output] boolean

true if the person is adult, false otherwise.