Answer of treeHeight
The height of a tree is the number of edges on the longest downward path between the root and a leaf.

You are given a tree with vertices numbered from 0 to N - 1. The tree is input as an array parent where parent[i] is the parent of the ith node of the tree. The root of the tree is the vertex number 0 and it is its own parent.

Find the height of the given tree.

Example

For the tree [0, 0, 1, 4, 1], the output should be 3.

    [input] array.integer parent
        The array of parents.

    [output] integer