Answer of rounders
We want to turn the given integer into a round number (i.e. a number that has only one non-zero digit) using a tail rounding approach. This means that at each step we take the last non 0 digit of the number and round it to 0 or to 10. If it's less than 5 we round it to 0 if it's larger than or equal to 5 we round it to 10 (rounding to 10 means increasing the next significant digit by 1).

Example

15 -> 20
1234 -> 1230 -> 1200 -> 1000
1445 -> 1450 -> 1500 -> 2000

    [input] integer value
        positive integer

    [output] integer
        rounded number.