February Marathon
A new parking lot has opened in your city. You are interested in its rates, and here is what you found out:

    The first 30 minutes (inclusive) of parking are free;
    From 31 minutes to 2 hours each 10 minutes cost 1$;
    Each 10 minutes over 2 hours cost 2$.

As you are charged for each 10 minutes, you should round up the number of minutes spent in the parking lot to the nearest multiple of 10, i.e. if you've been in the parking lot for, as an example, 43 minutes, you should pay for 50 minutes.

You are given 2 strings: the time you typically arrive at your usual parking lot and the time you typically leave it. Now you want to know how much you'll have to pay in the new one.

It's guaranteed that all events take place in one day.

Example

    For timeIn = "14:11" and timeOut = "14:39" the output should be 0.
    You usually spend 28 minutes in the parking lot, so you won't have to pay anything.

    For timeIn = "17:52" and timeOut = "18:32" the output should be 1.
    You spend 40 minutes in the parking lot, so you'll need to pay for the first 10 minutes after half an hour.

    For timeIn = "12:20" and timeOut = "14:54" the output should be 17.
    You spend 2 hours and 34 minutes in the parking lot, so you'll need to pay 9$ for the first 2 hours (there are 90 minutes between 30 minutes and 2 hours, each 10 minutes of which cost 1$) and 4 * 2 = 8$ for the next 34 minutes, which makes a total of 8 + 9 = 17$.

    [input] string timeIn
        The time you usually arrive at the parking lot in the format hh:mm.

    [input] string timeOut
        The time you usually leave the parking lot in the format hh:mm.

    [output] integer
        The number of dollars you'll need to pay.