February Marathon
Let's play Tetris! At first let's define the rules, especially since they probably differ from those you've heard before.

There is an empty field with 20 rows and 10 columns, which is initially empty. Random pieces appear on the field, each composed of four square blocks. You can't change the piece's shape, but you can rotate it 90 degree clockwise (possibly several times) and choose within which columns it will appear. Once you've rotated the piece and have set its starting position, it appears at the topmost row where you decided to put it, and falls down until it can't fall any further. The objective of the game is to create horizontal lines of 10 blocks. When such a line is created, it disappears, and all lines above the deleted one move down. For each deleted row the player receives 1 point.

Your task is to implement the algorithm that places each new piece optimally. The piece is considered to be placed optimally, if:

    the total number of blocks in the rows this piece will occupy after falling down is maximized;
    among all positions with that value maximized, this position requires the least number of rotations;
    among all positions that require the minimum number of rotations, this one is the leftmost one (i.e. the leftmost block's position is as far to the left as possible).

The piece can't leave the field. It is guaranteed that it is always possible to place the piece in the field.

Implement this algorithm and calculate the number of points you will get for the given set of pieces.

Example

For

pieces = [[[".", "#", "."], 
           ["#", "#", "#"]],
          [["#", ".", "."], 
           ["#", "#", "#"]],
          [["#", "#", "."], 
           [".", "#", "#"]],
          [["#", "#", "#", "#"]],
          [["#", "#", "#", "#"]],
          [["#", "#"], 
           ["#", "#"]]]

the output should be
tetrisGame(pieces) = 1.

For the explanation let's represent each block by the index of the piece it belongs to.
After the first 5 blocks the field will look like this:

...
. . . . . . . . . .
. . . . . . . . . .
. . . . . . . . . .
. . . . . . . . . .
. . . . . . . . 3 4
. . . . . . . . 3 4
. 0 . 1 . 2 2 . 3 4
0 0 0 1 1 1 2 2 3 4

Note that the first 3 pieces fell down without any rotation, while the 4th and the 5th pieces were rotated one time each.

Since there's a row of 10 blocks, it is deleted, and you get 1 point.
Afterward, the last piece falls down and the final field looks like this:

...
. . . . . . . . . .
. . . . . . . . . .
. . . . . . . . . .
. . . . . . . . . .
. . . . . . . . . .
5 5 . . . . . . 3 4
5 5 . . . . . . 3 4
. 0 . 1 . 2 2 . 3 4

    [input] array.array.array.char pieces
        A non-empty array of pieces in the order they fall. Each piece is represented as a rectangular matrix, where '#' represents a block and '.' represents an empty cell.

Each piece consists of 4 blocks, and each block shares a common side with at least one another block. It's guaranteed that each piece contains neither empty rows nor empty columns.

    [output] integer
        The number of points you will have by the end of the game.