Create isDivisibleBy6.java
A masked number is a string that consists of digits and one asterisk (*) that should be replaced by exactly one digit. Given a masked number find all the possible options to replace the asterisk with a digit to produce an integer divisible by 6.

Example

for '1*0' output should be ['120', '150', '180']

[input] string inputString

a masked number
[output] array.string

sorted array of strings representing all non-negative integers that correspond to the given mask and are divisible by 6.