Answer of toAndFro
A person is moving along a straight line. Initially he is at point A. He goes to point B from A with speed equal to 1 meter per second. Immediately after reaching B he turns around and heads to A from B with the same speed. After reaching point A he turns around once again and heads to B. etc. We need an algorithm that identified the location of the person at any given moment in time.

It's guaranteed that A and B are two different points.

Example

toAndFroNaive(2, 4, 3) = 3

    [input] integer a
        coordinate of point A (in meters)

    [input] integer b
        coordinate of point B (in meters)

    [input] integer t
        positive integer representing time (in seconds)

    [output] integer
        coordinate of the person t seconds after he left A