Answer of createAnagram
P string x is an anagram of a string y if one can get y by rearranging the letters of x. For example, the strings "MITE" and "TIME" are anagrams, so are "BABA" and "AABB", but "ABBAC" and "CAABA" are not.

You are given two strings s and t of the same length, consisting of uppercase English letters. Your task is to find the minimum number of "replacement operations" needed to get some anagram of the string t from the string s. A replacement operation is performed by picking exactly one character from the string s and replacing it by some other character.

Example

    createAnagram("AABAA", "BBAAA") = 1
    createAnagram("OVGHK", "RPGUC") = 4

    [input] string s

    [input] string t

    [output] integer
        The minimum number of replacement operations needed to get an anagram of the string t from the string s.