Answer of CakewalkSimulator
In a magical land, there is an old witch living in an old hut. She wants to buy stuff in the city, but she's ancient and cannot walk well. Because of that, the witch sends a magic cake which can move in any direction and bring stuff to her on a trip.

The witch will tell the cake the number of steps and direction, which can be "right", "down", "downright" or "downleft". If the number of steps is negative, the direction changes to the opposite, i.e. to "left", "up", "upleft" or "upright" respectively.

Your task is to draw a map for the cake of the smallest size possible. The map should consist of 0 at positions the cake won't visit, and numbers indicating the cake's steps starting from 1.

Example:

    For steps = 3 and direction = "downright"
    the answer should be

CakewalkSimulator(steps, direction) = [[1,0,0],
                                       [0,2,0],
                                       [0,0,3]]

    For steps = -5 and direction = "right"
    the answer should be

CakewalkSimulator(steps, direction) = [[5, 4, 3, 2, 1]]

Since steps < 0, the direction is actually "left".

    [input] integer steps
        The number of steps, abs(steps) > 0.

    [input] string direction
        The direction the cake should go to: "right", "downright", "down" or "downleft".

    [output] array.array.integer
        The map for the cake of the minimal size.