Create BigNumber.java
Big Number Calculation

Your task is to perform either addition or subtraction of two huge non-negative numbers stored as strings. Each number can be either float or integer, and may or may not contain the decimal point ('.'). If the decimal point is present, there's always at least one digit before it.

Apply the operator to value1 and value2 and return the answer as a string without trailing zeros.

If the answer is integer, omit the decimal point.

Example:

BigNumber("9.2", "3.42", "+") = "12.62"

[input] string value1

The first number.
[input] string value2

The second number.
It is guaranteed that if operator = '-', then value2 ≤ value1.
[input] string operator

The operator to apply, either '-' or '+'.
[output] string

The exact result as a string.