Create parameter.java
Let's define a parameter of number n as the least common multiple (LCM) of the sum of its digits and their product.

Calculate the parameter of the given number n.

Example:

parameter(1) = 1
parameter(22) = 4
Both the sum and the product of digits equal 4, and LCM(4, 4) = 4.
[input] integer n

A positive integer.
[output] integer

The parameter of the given number. It is guaranteed that the answer fits in a 32-bit integer.