Update josephusProblem.java
magine the following situation for a given integers n and k. There are n people standing in a circle. They are numbered from 1 through n in clockwise direction. The counting out begins at person #1 and continues around the circle in a clockwise direction. In each step, k-1 people are skipped and the next person is removed from the circle. The elimination proceeds around the circle (which is becoming smaller and smaller as people get removed), until only one person remains, who is announced a winner.

The task is to find the place in the initial circle that would guarantee a win.

Example

for n = 3, k = 2 the answer should be 3

[input] integer n

positive integer representing the number of people
[input] integer k

positive integer
[output] integer

index of the winning place