Create blackAndWhiteBoard.java
Consider a rectangular board consisting of h * w cells, which are initially white. You are to process several queries of the following types:

"x a b" - color the white cell (a, b) (0-based coordinates, the first one is a row index, and the second one is a column index) black;
"> a b" - find the leftmost white cell to the right of the white cell (a, b);
"< a b" - find the rightmost white cell to the left of the white cell (a, b);
"v a b" - the same, but the search should be done downwards;
"^ a b" - the same, but the search should be done upwards;
For each query, except the ones of the first type, find the answer.

Example

For h = 3, w = 5, queries = ["v 1 2", "x 2 2", "v 1 2", "> 2 1", "x 2 3", "> 2 1", "< 2 0"], the output should be [[2, 2], [-1, -1], [2, 3], [2, 4], [-1, -1]]

[input] integer h

A positive integer.
[input] integer w

A positive integer.
[input] array.string queries

Queries in the above-described format.
[output] array.array.integer

For each query except the ones of the first type, store the answer's coordinates in the array. If the desired cell doesn't exist, store [-1, -1] instead. The answers should be stored in the same order as the queries.