Create factorability.java
Given coefficients a, b, and c, determine whether the quadratic expression ax2 + bx + c is factorable over the integers.

A quadratic expression is considered factorable over the integers if there exist two numbers whose sum is equal to b and whose product is equal to a * c. In other words, if there are two numbers m and n such that m + n = b and m * n = a * c, then the expression is factorable.

Example

factorability(1, 5, 6) = true
3 + 2 = 5 = b and 3 * 2 = 6 = a * c
factorability(3, 5, 1) = false
[input] integer a

Coefficient of the x2 term (a non-zero integer).
[input] integer b

Coefficient of the x term (an integer which might be zero).
[input] integer c

Constant term (the term without an x value; also an integer which could be zero).
[output] boolean

Return true if the expression is factorable, and false if it isn't.